+++ changes.md
@@ -10,6 +10,7 @@
 - For now use `distance` and `comment` instead of `distance_field` and `comment_field` in NewickSerializer.
 - Changed default format of `to_dict` to contain `identifier`.
 - Rewrite RowSerializer
+- Refactor RelationSerializer
 
 ## Version 0.2.3 ##
 - Streamline `DotExporter`.